% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes.R
\name{add_covariates}
\alias{add_covariates}
\title{#' @export
align_to <- function(x, group_info) {
  #library(furrr)
  #albrowser()
  x %>% unnest(cols = c(neurons)) %>%
    mutate(across(starts_with("AD"), ~map2(.x, shift, ~.x - .y))) %>%
    select(-shift) %>%
    pivot_longer(!counter_total_trials, values_to = "times")
    #nest(neurons = starts_with("AD")) # re-nest?
}}
\usage{
add_covariates(x, group_info, y)
}
\description{
#' @export
align_to <- function(x, group_info) {
  #library(furrr)
  #albrowser()
  x %>% unnest(cols = c(neurons)) %>%
    mutate(across(starts_with("AD"), ~map2(.x, shift, ~.x - .y))) %>%
    select(-shift) %>%
    pivot_longer(!counter_total_trials, values_to = "times")
    #nest(neurons = starts_with("AD")) # re-nest?
}
}
